// cs631-db2/src/app/rest/model/student.ts
/**
 * CS 631 Project
 * Roderic Henry & Amruta Abhyankar
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Department } from './department';

export interface Student { 
    address?: string;
    department?: Department;
    email?: string;
    firstname?: string;
    highSchool?: string;
    id?: number;
    lastname?: string;
    major?: string;
    phone?: string;
    year?: number;
}