// cs631-db2/src/app/rest/api/sectionController.service.ts
export class SectionControllerService {

    protected basePath = '//localhost:8080/api';
    public defaultHeaders = new HttpHeaders();
    public configuration = new Configuration();

    constructor(protected httpClient: HttpClient, @Optional()@Inject(BASE_PATH) basePath: string, @Optional() configuration: Configuration) {
        if (basePath) {
            this.basePath = basePath;
        }
        if (configuration) {
            this.configuration = configuration;
            this.basePath = basePath || configuration.basePath || this.basePath;
        }
    }

    /**
     * @param consumes string[] mime-types
     * @return true: consumes contains 'multipart/form-data', false: otherwise
     */
    private canConsumeForm(consumes: string[]): boolean {
        const form = 'multipart/form-data';
        for (const consume of consumes) {
            if (form === consume) {
                return true;
            }
        }
        return false;
    }


    /**
     * index
     * 
     * @param course course
     * @param semester semester
     * @param year year
     * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
     * @param reportProgress flag to report request and response progress.
     */
    public indexUsingGET2(course?: string, semester?: string, year?: number, observe?: 'body', reportProgress?: boolean): Observable<Array<Section>>;
    public indexUsingGET2(course?: string, semester?: string, year?: number, observe?: 'response', reportProgress?: boolean): Observable<HttpResponse<Array<Section>>>;
    public indexUsingGET2(course?: string, semester?: string, year?: number, observe?: 'events', reportProgress?: boolean): Observable<HttpEvent<Array<Section>>>;
    public indexUsingGET2(course?: string, semester?: string, year?: number, observe: any = 'body', reportProgress: boolean = false ): Observable<any> {




        let queryParameters = new HttpParams({encoder: new CustomHttpUrlEncodingCodec()});
        if (course !== undefined && course !== null) {
            queryParameters = queryParameters.set('course', <any>course);
        }
        if (semester !== undefined && semester !== null) {
            queryParameters = queryParameters.set('semester', <any>semester);
        }
        if (year !== undefined && year !== null) {
            queryParameters = queryParameters.set('year', <any>year);
        }

        let headers = this.defaultHeaders;

        // to determine the Accept header
        let httpHeaderAccepts: string[] = [
            '*/*'
        ];
        const httpHeaderAcceptSelected: string | undefined = this.configuration.selectHeaderAccept(httpHeaderAccepts);
        if (httpHeaderAcceptSelected != undefined) {
            headers = headers.set('Accept', httpHeaderAcceptSelected);
        }

        // to determine the Content-Type header
        const consumes: string[] = [
        ];

        return this.httpClient.request<Array<Section>>('get',`${this.basePath}/sections`,
            {
                params: queryParameters,
                withCredentials: this.configuration.withCredentials,
                headers: headers,
                observe: observe,
                reportProgress: reportProgress
            }
        );
    }

}